# Faster sax