Optimizations and a couple tricks to get this to be linear in size of sequence:

```python
In [3]: %timeit -r1 find_discords_hotsax(np.random.random(5000), num_discords=2)
8.85 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)

In [4]: %timeit -r1 find_discords_hotsax(np.random.random(10000), num_discords=2)
17.8 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)

In [5]: %timeit -r1 find_discords_hotsax(np.random.random(20000), num_discords=2)
36.1 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)

In [6]: %timeit -r1 find_discords_hotsax(np.random.random(40000), num_discords=2)
1min 18s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
```

Tricks: 

- precompute all znorms (old code does a lot of redundant calculations)
- don't search over most common PAA sequence (in sequences with no anomalies, these will dominate by far, and are the least interesting regions)
- limit random search.  as noted in the paper, random search is O(n2)